_lastProgress moved to Transform2DEffect class
